==> code/hyr/5_model_stacking.ipynb <==
import pickle
import lightgbm as lgb
import os
import pandas as pd
import numpy as np
from sklearn.model_selection import KFold

FILE_DIR = 'sub_model_output'
DATA_PATH = '.'

# Read

df_logit_list = []
for name in os.listdir(FILE_DIR):
    if 'mask' in name or 'ipynb_checkpoints' in name:
        continue
    df_logit = pickle.load(open('%s/%s'%(FILE_DIR,name), 'rb'))
    df_logit_list.append(df_logit)
df_train_origin_weibo = pd.read_csv('%s/train.origin_weibo.csv'%DATA_PATH, sep='\t')
df_train_origin_weibo.index = df_train_origin_weibo.WeiboId

for i in range(len(df_logit_list)):
    print(sum(df_logit_list[0]['WeiboId'] == df_logit_list[i]['WeiboId']))

def precision_score_cch(predictions, ground_truths):
    
    predictions, ground_truths = np.array(predictions)-1, np.array(ground_truths)-1
    y_pred = predictions
    y_true = ground_truths

    w=[1,10,50,100,300]
    n = len(y_true)
    count_r = [0 for i in range(5)]
    count = [0 for i in range(5)]
    for i in range(n):
        count[y_true[i]] += 1
        if y_pred[i] == y_true[i]:
            count_r[y_pred[i]] += 1
    sum1 = sum(w[i]*count_r[i] for i in range(5))
    sum2 = sum(w[i]*count[i] for i in range(5))
    precision = sum1/sum2
    return precision

def df_logit2score(df_logit, ground_truths):
    logit_name = sorted(list(set(df_logit.columns)-set(['WeiboId'])))
    logit = np.array(df_logit[logit_name])
    if len(logit_name) == 5:
        predictions = np.argmax(np.array(logit), 1)
    elif len(logit_name) == 1:
        predictions = np.array([count2idx(num) for num in logit[:, 0]])
    else:
        return -1
    cnt = 0
    for p, g in zip(predictions, ground_truths):
        if p+1 == g:
            cnt += 1
    return precision_score_cch(predictions+1, ground_truths), cnt/predictions.shape[0]

def count2idx(num):
    if 0 <= num <= 10:
        return 0
    elif 10 <= num <= 50:
        return 1
    elif 50 <= num <= 150:
        return 2
    elif 150 <= num <= 300:
        return 3
    elif num > 300:
        return 4
    return 0
    
df_train_origin_weibo = pd.read_csv('%s/train.origin_weibo.csv'%DATA_PATH, sep='\t')
df_train_origin_weibo.index = df_train_origin_weibo.WeiboId


group_truths = np.array(df_train_origin_weibo['ForwordCount'].apply(count2idx)) + 1
df_logit_list = []
feature_name_list = []
for name in os.listdir(FILE_DIR):
    if 'mask' in name or 'ipynb_checkpoints' in name:
        continue

    df_logit = pickle.load(open('%s/%s'%(FILE_DIR,name), 'rb'))
    logit_name = list(set(df_logit.columns)-set(['WeiboId']))
    df_logit = df_logit[logit_name]
    logit_name = ['%s_%s'%(name.split('.')[0], f) for f in logit_name]
    df_logit.columns = logit_name
    feature_name_list.extend(logit_name)
    df_logit_list.append(df_logit)
    print(name, df_logit2score(df_logit, group_truths))

# df_LSTM_yjq = pickle.load(open('%s/df_LSTM_yjq.pickle'%(FILE_DIR), 'rb'))
# df_LSTM_yjq = pd.concat([df_LSTM_yjq[['WeiboId']],pd.get_dummies(df_LSTM_yjq['LSTM_pre'])], 1)
# pickle.dump(df_LSTM_yjq, open('%s/df_LSTM_yjq.pickle'%(FILE_DIR), 'wb'))

# pickle.load(open('%s/cch_lgbReg.pickle'%(FILE_DIR), 'rb'))['lgbReg_cch_fowardcount'].apply(count2idx).value_counts(1)

df_logits = pd.concat(df_logit_list, 1)
df_logits.index = pickle.load(open('sub_model_output/df_lgb.pickle', 'rb')).WeiboId
feature_name_list = list(df_logits.columns)
df_logits['label'] = df_train_origin_weibo['ForwordCount'].apply(count2idx)
df_logits['type'] = ['train'] * 18000 + ['test'] * 2329

def get_train_test_from_df(df, feature_name_list):
    df_train = df.iloc[:18000]
    df_test = df.iloc[18000:]
    train_x_fix, train_y, test_x_fix = \
        np.array(df_train[feature_name_list]), np.array(df_train['label']), np.array(df_test[feature_name_list])
    
        
    return train_x_fix, train_y, test_x_fix

train_x, train_y, test_x = \
get_train_test_from_df(df_logits, feature_name_list)

train_x.shape, train_y.shape, test_x.shape

# 权重赋予

from sklearn.metrics import roc_auc_score

def fit_binary_model(train_x, train_y):
    trn_data = lgb.Dataset(train_x, train_y)
    clf = lgb.train({  
                'boosting_type': 'gbdt',  
                'objective': 'binary',  
                'metric': 'auc',  
                'num_leaves': 8,  
                'max_depth': 3,
                'min_data_in_leaf': 100,  
                'learning_rate': 0.06,  
                'feature_fraction': 0.8,  
                'bagging_fraction': 0.8,  
                'bagging_freq': 5,  
                'lambda_l1': 0.5,  
                'lambda_l2': 0.5,  
                'min_gain_to_split': 0.2,  
                'verbose': -1, 
                }, 
                trn_data, 
                num_boost_round = 100,
                valid_sets = [trn_data], 
                verbose_eval = 100, 
                early_stopping_rounds = 100)
    return clf

def eval_logit(logit, label):
    return roc_auc_score(label, logit)
    
def cross_validation_fit_weight(train_x, train_y):
    n_flod = 5
    folds = KFold(n_splits=n_flod, shuffle=True)
    train_x = np.array(train_x)
    train_y = np.array(train_y)
    score_train = np.zeros((len(train_x), ))
    for n_fold, (trn_idx, val_idx) in enumerate(folds.split(train_x, train_y)):
        
        trn_x, trn_labels = train_x[trn_idx], train_y[trn_idx]
        val_x, val_labels = train_x[val_idx], train_y[val_idx]
        model = fit_binary_model(trn_x, trn_labels)
        score_train[val_idx] = model.predict(val_x)
        
        print(eval_logit(model.predict(trn_x), trn_labels))
        print(eval_logit(score_train[val_idx], val_labels))
    
    print(eval_logit(score_train, train_y))
    return score_train

def fit_train_test_weight(train_x, test_x):
    train_y = np.array([0]*train_x.shape[0] + [1]*test_x.shape[0])
    return cross_validation_fit_weight(np.concatenate([train_x, test_x]), train_y)


dataset_weights = fit_train_test_weight(train_x, test_x)

# 训练

SEED = 666 
# import random
# SEED = random.randint(0, 1314)
# print(SEED)

params = {  
    'boosting_type': 'gbdt',  
    'objective': 'multiclass',  
    'num_class': 5,  
    'metric': 'multi_error',  
    'num_leaves': 16,  
    'max_depth': 4,
    'min_data_in_leaf': 100,  
    'learning_rate': 0.005,  
    'feature_fraction': 0.8,  
    'bagging_fraction': 0.8,  
    'bagging_freq': 5,  
    'lambda_l1': 0.5,  
    'lambda_l2': 0.5,  
    'min_gain_to_split': 0.2,  
    'verbose': -1, 
    'feature_fraction_seed':SEED,
    'bagging_seed':SEED,

} 

def precision_score_cch(predictions, ground_truths):
    
    predictions, ground_truths = np.array(predictions)-1, np.array(ground_truths)-1
    y_pred = predictions
    y_true = ground_truths

    w=[1,10,50,100,300]
    n = len(y_true)
    count_r = [0 for i in range(5)]
    count = [0 for i in range(5)]
    for i in range(n):
        count[y_true[i]] += 1
        if y_pred[i] == y_true[i]:
            count_r[y_pred[i]] += 1
    sum1 = sum(w[i]*count_r[i] for i in range(5))
    sum2 = sum(w[i]*count[i] for i in range(5))
    precision = sum1/sum2
    return precision

def fit_lgb(train_x, train_y, sub_dataset_weights):
    class_weights = [1,10,50,100,300]
#     sub_dataset_weights = [1] * len(sub_dataset_weights)
    trn_data = lgb.Dataset(train_x, train_y,
                           weight=[class_weights[int(y)]*sub_dataset_weights[i] for i, y in enumerate(train_y)])
    clf = lgb.train(params, 
                    trn_data, 
                    num_boost_round = 400,
                    valid_sets = [trn_data], 
                    verbose_eval = 100, 
                    early_stopping_rounds = 100
                   )
    return clf

def eval_logit(logit, label):
    prediction = np.argmax(logit, 1)
    return precision_score_cch(prediction+1, np.array(label).astype('int')+1)
    
def cross_validation_lgb(train_x, train_y, test_x, dataset_weights):
    n_flod = 25
    folds = KFold(n_splits=n_flod, shuffle=True, random_state=SEED)
    train_x = np.array(train_x)
    train_y = np.array(train_y)
    score_train = np.zeros((len(train_x), 5))
    score_test = np.zeros((len(test_x), 5))
    train_score_list = []
    for flod, (trn_idx, val_idx) in enumerate(folds.split(train_x, train_y)):
        
        trn_x, trn_labels = train_x[trn_idx], train_y[trn_idx]
        val_x, val_labels = train_x[val_idx], train_y[val_idx]
        model = fit_lgb(trn_x, trn_labels, dataset_weights[trn_idx])
        score_train[val_idx] = model.predict(val_x)
        score_test += model.predict(test_x)/n_flod
        
        train_score = eval_logit(model.predict(trn_x), trn_labels)
        print(flod)
        print(train_score)
        train_score_list.append(train_score)
        print(eval_logit(score_train[val_idx], val_labels))
    
    prediction = np.argmax(score_train, 1)
    val_score = precision_score_cch(prediction+1, np.array(train_y).astype('int')+1)
    
    print('train score %f, val score %f'%(np.mean(train_score_list), val_score))

    return score_train, score_test

score_train, score_test = cross_validation_lgb(train_x, train_y, test_x, dataset_weights)

# sub_feature_name = ['lgb_hyr_0', 'lgb_hyr_1', 'lgb_hyr_2', 'lgb_hyr_3','lgb_hyr_4']
# sub_feature_name =  ['nn_hyr_loss_ce_0', 'nn_hyr_loss_ce_1', 'nn_hyr_loss_ce_2', 'nn_hyr_loss_ce_3', 'nn_hyr_loss_ce_4']
    
# prediction = np.argmax(np.array(df_logits[sub_feature_name]), 1)
prediction = np.argmax(score_train, 1)
print(pd.Series(prediction).value_counts(1))
print('baseline', precision_score_cch(prediction+1, np.array(train_y).astype('int')+1))

prediciton_test = np.argmax(score_test, 1)
df_submit = pd.DataFrame({
    'WeiboId':list(df_logits.query('type=="test"').index),
    'ForwardScale':prediciton_test+1
})
SUBMIT_PATH = '.'
df_submit.to_csv('%s/submission.csv'%SUBMIT_PATH, sep='\t', index=False)

df_submit = pd.read_csv('%s/submission.csv'%SUBMIT_PATH,sep='\t')

df_submit_old = pd.read_csv('submit/submission_12_12.csv', sep='\t')

cnt = 0
for a, b in zip(df_submit['ForwardScale'], df_submit_old['ForwardScale']):
    if a==b:
        cnt += 1
        
print(cnt/2329)
